added filters in config
