+++ lib/login.js
@@ -3,6 +3,7 @@ const Messenger = require('facebook-chat-api')
 const fs = require('fs')
 const sendError = require('./error.js')
 const getChannel = require('./getChannel.js')
+const getEnvConfig = require('./getEnvConfig.js')
 const discord = new Discord.Client()
 var loggedin = false
 
@@ -76,27 +77,5 @@ function getConfig () {
 
   // added for backwards compatibility, discord.showUsername is deprecated
   if (!config.facebook.showUsername) config.facebook.showUsername = config.discord.showUsername
-  return config
-}
-
-function getEnvConfig () {
-  var config = {
-    facebook: {
-      username: process.env.FACEBOOK_USERNAME,
-      password: process.env.FACEBOOK_PASSWORD,
-      forceLogin: process.env.FACEBOOK_FORCE_LOGIN,
-      showUsername: process.env.FACEBOOK_SHOW_USERNAME,
-      boldUsername: process.env.FACEBOOK_BOLD_USERNAME
-    },
-    discord: {
-      token: process.env.DISCORD_TOKEN,
-      guild: process.env.DISCORD_GUILD,
-      category: process.env.DISCORD_CATEGORY,
-      sendNotifications: process.env.DISCORD_SEND_NOTIFICATIONS
-    }
-  }
-  if (!config.discord.token || !config.facebook.username || !config.facebook.password) {
-    sendError('Failed to get config information')
-  }
   return config
 }
